+++ QCloudAPI_SDK/QCloudAPI_SDK/Module/Batch.cs
@@ -1,0 +1,10 @@
+namespace QCloudAPI_SDK.Module
+{
+    public class Batch : Base
+    {
+        public Batch()
+        {
+            serverHost = "batch.api.qcloud.com";
+        }
+    }
+}